+++ CHANGELOG.md
@@ -7,7 +7,10 @@ All notable changes to this project will be documented in this file. This projec
 
 ### Added
 
-- **BREAKING** Made the followign changes to the `Contracts\Query\QueryParameters` interface:
+- **BREAKING** The `Contracts\Schema\Schema` interface now has `isFilter()`, `isSparseField` and `isSortable()` methods.
+  These methods have been added to the abstract schema class provided by this package, so this is unlikely to have a
+  significant impact on implementing packages.
+- **BREAKING** Made the following changes to the `Contracts\Query\QueryParameters` interface:
     - New `unrecognisedParameters` method. This returns any query parameters that are not defined by the JSON:API
       specification, which allows implementations to add support for additional query parameters as needed.
     - The `filters` method now returns a `FilterParameters` object or null. Previously it returned an array or null.